refactor: fix routing issus by changing module loading
When a page is refreshed in the browser e.g. /login
the path is stripped and the page loads with root /
The problem was, that the fallback route component
is already rendered before the other routes get rendered
because they are still waiting for some resources to load.
But with the fallback route already in place, it then
redirects to / before other routes are registerd.